refactoring SMTimeSeries into its own file
